Refactor task creation

+++ kernel/src/abi/string/strncpy_to_user.c
@@ -1,10 +1,10 @@
 #include <abi/string.h>
-#include <task.h>
+#include <mm_context.h>
 
 size_t strncpy_to_user(struct mm_context * mm_context, void * dest,
                        const char * src, size_t num)
 {
-    unsigned length = strlen(src) + 1;
+    size_t length = strlen(src) + 1;
 
     num = length < num ? length : num;
